requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==0.27.8
anthropic==0.7.1
google-generativeai==0.3.0
pydantic==2.0.3
redis==5.0.0
sqlalchemy==2.0.21
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
